# aesni: port to intrinsics