--- a/library/aesni.h
+++ b/library/aesni.h
@@ -32,9 +32,10 @@
 #define MBEDTLS_AESNI_AES      0x02000000u
 #define MBEDTLS_AESNI_CLMUL    0x00000002u
 
-#if defined(MBEDTLS_HAVE_ASM) && defined(__GNUC__) &&  \
-    ( defined(__amd64__) || defined(__x86_64__) )   &&  \
-    ! defined(MBEDTLS_HAVE_X86_64)
+// TODO figure out proper test to enable
+#if defined(MBEDTLS_HAVE_ASM) && \
+    ( defined(__amd64__) || defined(__x86_64__) || defined(_M_X64) ) && \
+    !defined(MBEDTLS_HAVE_X86_64)
 #define MBEDTLS_HAVE_X86_64
 #endif