… library/aesni.c