## Description
This moves the existing AES-NI inline assembly to intrinsics. I've tested the code on msvc and at least checked it compiled with gcc, and expect it to compile with clang and maybe icc. I suspect some small changes to the generated code are improvements over the inline asm (normal things like better register allocation, aligning jump targets, inlining, etc., which are missed from the asm implementation), but in any case having a AES-NI implementation at all is a large improvement for the msvc build.

For the time being I've just skipped the GCM code as I've seen there is another active PR which is modifying it.


## Status
IN DEVELOPMENT

## Requires Backporting
NO

## Migrations
NO

## Additional comments
I am submitting this for comments and to hopefully get it merged into mainline after taking feedback. The performance difference is rather large so I would very much like to have it "just work" on msvc without carrying a patch around forever.

This PR on amd 3990x (zen2):
```
>benchmark.exe aes_cbc aes_ccm aes_xts

  AES-CBC-128              :     673579 KiB/s
  AES-CBC-192              :     623720 KiB/s
  AES-CBC-256              :     590403 KiB/s
  AES-XTS-128              :     599192 KiB/s
  AES-XTS-256              :     529954 KiB/s
  AES-CCM-128              :     326978 KiB/s
  AES-CCM-192              :     314479 KiB/s
  AES-CCM-256              :     300769 KiB/s
```

current development branch:
```
>benchmark.exe aes_cbc aes_ccm aes_xts

  AES-CBC-128              :     181729 KiB/s
  AES-CBC-192              :     156952 KiB/s
  AES-CBC-256              :     139389 KiB/s
  AES-XTS-128              :     164933 KiB/s
  AES-XTS-256              :     128241 KiB/s
  AES-CCM-128              :      81466 KiB/s
  AES-CCM-192              :      70942 KiB/s
  AES-CCM-256              :      63345 KiB/s
```

## Todos
??
- [ ] Tests
- [ ] Documentation
- [ ] Changelog updated
- [ ] Backported