@shuffle2 2022-06-23
> aesni.c: In function ‘write_roundkey_192’:
aesni.c:301:5: warning: implicit declaration of function ‘_mm_storeu_si64’ [-Wimplicit-function-declaration]
     _mm_storeu_si64(d, rk_hi);

This [appears to be because](https://github.com/mortie/swaylock-effects/issues/44) you're using an ancient version of gcc:  
`gcc: /usr/bin/gcc : gcc (Ubuntu 5.4.0-6ubuntu1~16.04.12) 5.4.0 20160609 `