@gilles-peskine-arm 2022-06-23
It's good to have intrinsics, but we can't break toolchains that currently work. This means we'd have to have two versions, which is unfortunate, but I think unavoidable. We do need to make sure both versions are tested in our CI, which will be the case as long as we keep testing with older gcc.